Added code to account for uncovered files
Since I have to instrument code manually I forgot to add in the files that don't have tests associated with them yet. If a user then tries to add a new test file and adds tests, the coverage would drop because the new lines etc. were not previously accounted for.
I've updated the coverage thresholds to account for the lowered coverage b because it's now reporting (correctly) that there are a couple files that don't have tests yet.